last update, stable

+++ api/app/models.py
@@ -1,12 +1,11 @@
 # app/models.py
 
-from sqlalchemy import Column, Integer, String, JSON, ForeignKey, DateTime, Float, Table, Boolean
+from sqlalchemy import Column, Integer, String, JSON, ForeignKey, DateTime, Float, Boolean
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base
 from geoalchemy2 import Geometry
 from sqlalchemy.sql import func
 from sqlalchemy.schema import UniqueConstraint
-from api.app.enums import IndustryEnum, SpecialtyEnum, WorkforceTypeEnum
 
 Base = declarative_base()
 
@@ -26,9 +25,9 @@ class Project(Base):
     industry_name = Column(String, nullable=True)  # Дополнительное название индустрии
     n_hours = Column(Float, default=0, nullable=True)  # Дополнительное поле часов
     workforce_type = Column(String, nullable=True)
-    specialists_total = Column(Float, nullable=True)
-    graduates_total = Column(Float, nullable=True)
-    all_total = Column(Float, nullable=True)
+    specialists_total = Column(Integer, nullable=True)
+    graduates_total = Column(Integer, nullable=True)
+    all_total = Column(Integer, nullable=True)
     metric_bool = Column(Boolean, nullable=True)
     metric_float = Column(Float, nullable=True)
